[feat] 체스보드 다시시작하는 기능 추가
